Improve install dotfiles on mingw.

+++ installdotfiles.sh
@@ -5,11 +5,11 @@ DOTFILESDIRREL=$(dirname $0)/dotfiles
 cd $DOTFILESDIRREL
 DOTFILESDIR=$(pwd -P)
 for DOTFILE in *; do
-	rm -iv "$HOME/.$DOTFILE"
-	if [[ $TERM != "cygwin" ]]
+	if [[ $(uname -s) != MINGW* ]]
 	then
+		rm -iv "$HOME/.$DOTFILE"
 		ln -sv "$DOTFILESDIR/$DOTFILE" "$HOME/.$DOTFILE"
 	else
-		cp -v "$DOTFILESDIR/$DOTFILE" "$HOME/.$DOTFILE"
+		cp -iv "$DOTFILESDIR/$DOTFILE" "$HOME/.$DOTFILE"
 	fi
 done
